Made designs, html generator, and data table.
I made a design for the playlist and album buttons. I made a data table with album information including the name, artist, and image file. I imported these files into a javascript file and I made a generator which generates the design for the album or playlist depending on the information given. It helped me be more efficient instead of manually having to make the designs for playlists.